Visual Update 1
Animations for the RA enemy.
Animations for the Skunk enemy.
Updated animations for the Boh Boy.
Added the Car as a static object.
Added the Recycle Bin as a static object.
Added the Picnic Bench as a static object.
New grass, dirt, brick, and road tiles.
More intense drunkenness effect on screen rotation.
Fixed jumping glitch.

Still need to update the GUI.
Need to iron out feature list for the week.