Big Physics Update 1
DID NOT MOVE TO NEWEST MAP; WAS OVERWORKED AS-IS.

-Boh Boy uses a momentum-driven movement system.  Will gather up speed
when running and skid to a stop.  Will lose momentum if he hits a wall.
Will lose vertical momentum if he bumps his head.

-Boh Boy can now double-jump.
-Boh Boy can now swim.
-Boh Boy can now perform wall-jumps.

Known issues:
-Boh Boy will wall-jump off of blocks that clearly aren't meant to be
used as such (like stepping stone blocks).
-Swimming consumes double jumps, meaning you can not jump out of water
and have to float out (can fix this)